pyproject.toml
[project]
name = "comitter"
version = "0.1.0"
description = ""
authors = [
    {name = "Rohan Ghiya",email = "rohanghiya@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitpython (>=3.1.45,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
